requirements.txt
# Python HTTP 서버용 의존성 (실제로는 Flask가 아닌 내장 HTTP 서버 사용)
# Railway 배포를 위한 최소 의존성

# OpenAI API
openai>=1.0.0
python-dotenv>=1.0.0

# QR Code 생성
qrcode[pil]>=7.4.2
Pillow>=10.0.0
